Restructure the src directory to appease my OCD about src directory structures.
