+++ src/components/post-meta-control.js
