--- src/index.js
@@ -1,4 +1,0 @@
-// Import the variation.
-import './variations';
-// Import our custom controls.
-import './controls';
